+++ src/cache/index.ts
@@ -15,6 +15,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
+import * as _ from 'lodash';
 import { normalizeString } from '../index';
 
 /**
@@ -35,11 +36,17 @@ export interface Cache {
      *
      * @param {any} key The key.
      * @param {any} value The value.
+     * @param {SetCacheValueOptions|null|undefined} [opts] Additional and custom options.
      *
      * @return {Promise<boolean>} The promise that indicates if operation was successful or not.
      */
-    set(key: any, value: any): Promise<boolean>;
+    set(key: any, value: any, opts?: SetCacheValueOptions | null | undefined): Promise<boolean>;
 }
+
+/**
+ * Options for 'Cache.set()' method.
+ */
+export type SetCacheValueOptions = { [key: string]: any };
 
 /**
  * A basic cache (client).
@@ -49,7 +56,7 @@ export abstract class CacheBase implements Cache {
     public async get<TValue = any, TDefault = TValue>(key: any, defaultValue?: TValue): Promise<TValue | TDefault> {
         try {
             return await this.getInner(normalizeCacheKey(key),
-                                       defaultValue);
+                defaultValue);
         } catch {
             return defaultValue;
         }
@@ -58,11 +65,41 @@ export abstract class CacheBase implements Cache {
     /** @inheritdoc */
     protected async abstract getInner(key: string, defaultValue: any): Promise<any>;
 
+    /**
+     * Tries to return a value from a key/value pair.
+     *
+     * @param {Object|undefined|null} opts The key/value pair.
+     * @param {any} key The key.
+     * @param {TDefault} [defaultValue] The custom default value.
+     *
+     * @return {TValue|TDefault} The value or the default value if not found.
+     */
+    protected getOptionValue<TValue = any, TDefault = TValue>(
+        opts: { [key: string]: any } | undefined | null, key: any,
+        defaultValue?: TDefault
+    ): TValue | TDefault {
+        key = normalizeString(key);
+
+        if (opts) {
+            for (const PROP in opts) {
+                if (normalizeString(PROP) === key) {
+                    return opts[PROP];
+                }
+            }
+        }
+
+        return defaultValue;
+    }
+
     /** @inheritdoc */
-    public async set(key: any, value: any): Promise<boolean> {
+    public async set(key: any, value: any, opts?: SetCacheValueOptions | null | undefined): Promise<boolean> {
+        if (_.isNil(opts)) {
+            opts = {} as any;
+        }
+
         try {
-            await this.setInner(normalizeCacheKey(key),
-                                value);
+            await this.setInner(normalizeCacheKey(key), value,
+                opts);
 
             return true;
         } catch {
@@ -70,8 +107,17 @@ export abstract class CacheBase implements Cache {
         }
     }
 
-    /** @inheritdoc */
-    protected async abstract setInner(key: string, defaultValue: any): Promise<void>;
+    /**
+     * The logic for 'set()' method.
+     *
+     * @param {string} key The key.
+     * @param {string} defaultValue The default value.
+     * @param {SetCacheValueOptions} opts Custom options.
+     */
+    protected async abstract setInner(
+        key: string, defaultValue: any,
+        opts: SetCacheValueOptions
+    ): Promise<void>;
 }
 
 function normalizeCacheKey(key: any): string {